requirements.txt
requests==2.32.3
pyqt6==6.8.0
pillow==11.1.0